deadunits/requirements.txt
absl-py>=0.7.0
gin-config>=0.1.2
mock
numpy>=1.15.4
tensor2tensor
tensorflow==1.14.0
six>=1.10
